HIP 15 | David | Refactoring patient controller test by extracting request builder from test class
